+++ equinox-shipping/Watchdog.Lambda/Function.fs
@@ -48,7 +48,7 @@ type Store internal (connector : DynamoStoreConnector, table, indexTable, cacheN
         let tailSleepInterval =         TimeSpan.FromMilliseconds 500.
         let statsInterval =             TimeSpan.FromMinutes 1.
         let streamsDop =                2
-        let loadMode =                  DynamoLoadModeConfig.Hydrate (context, streamsDop)
+        let loadMode =                  Propulsion.DynamoStore.WithData (streamsDop, context)
         Handler.Config.CreateDynamoSource(Log.Logger, sink, (indexClient, checkpoints, loadMode, fromTail, batchSizeCutoff, tailSleepInterval, statsInterval), trancheIds)
 
 /// Wiring for Source and Sink running the Watchdog.Handler