+++ propulsion-projector/SourceConfig.fs
@@ -14,7 +14,7 @@ type SourceConfig =
 #if dynamo    
     | Dynamo of indexStore : Equinox.DynamoStore.DynamoStoreClient
         * checkpoints : Propulsion.Feed.IFeedCheckpointStore
-        * loading : DynamoLoadModeConfig
+        * loading : Propulsion.DynamoStore.EventLoadMode
         * startFromTail : bool
         * batchSizeCutoff : int
         * tailSleepInterval : TimeSpan
@@ -23,7 +23,7 @@ type SourceConfig =
 #if esdb     
     | Esdb of client : EventStore.Client.EventStoreClient
         * checkpoints : Propulsion.Feed.IFeedCheckpointStore
-        * hydrateBodies : bool
+        * withData : bool
         * startFromTail : bool
         * batchSize : int
         * tailSleepInterval : TimeSpan
@@ -32,7 +32,7 @@ type SourceConfig =
 #if sss     
     | Sss of client : SqlStreamStore.IStreamStore
         * checkpoints : Propulsion.Feed.IFeedCheckpointStore
-        * hydrateBodies : bool
+        * withData : bool
         * startFromTail : bool
         * batchSize : int
         * tailSleepInterval : TimeSpan
@@ -43,18 +43,14 @@ and [<NoEquality; NoComparison>] CosmosFeedConfig =
     | Ephemeral of processorName : string
     | Persistent of processorName : string * startFromTail : bool * maxItems : int option * lagFrequency : TimeSpan
 // #endif
-#if dynamo     
-and [<NoEquality; NoComparison>] DynamoLoadModeConfig =
-    | Hydrate of monitoredContext : Equinox.DynamoStore.DynamoStoreContext * hydrationConcurrency : int
-#endif
 
 module SourceConfig =
 // #if cosmos    
     module Cosmos =
         open Propulsion.CosmosStore
-        let start log (sink : Propulsion.Streams.Default.Sink) categoryFilter
+        let start log (sink : Propulsion.Streams.Default.Sink) categories
             (monitoredContainer, leasesContainer, checkpointConfig, tailSleepInterval) : Propulsion.Pipeline * (TimeSpan -> Task<unit>) option =
-            let parseFeedDoc = EquinoxSystemTextJsonParser.enumStreamEvents categoryFilter
+            let parseFeedDoc = EquinoxSystemTextJsonParser.enumStreamEvents (fun x -> Array.contains x categories)
             let observer = CosmosStoreSource.CreateObserver(log, sink.StartIngester, Seq.collect parseFeedDoc)
             let source =
                 match checkpointConfig with
@@ -74,16 +70,13 @@ module SourceConfig =
 #if dynamo    
     module Dynamo =
         open Propulsion.DynamoStore
-        let start (log, storeLog) (sink : Propulsion.Streams.Default.Sink) categoryFilter
-            (indexStore, checkpoints, loadModeConfig, startFromTail, batchSizeCutoff, tailSleepInterval, statsInterval) : Propulsion.Pipeline * (TimeSpan -> Task<unit>) option =
-            let loadMode =
-                match loadModeConfig with
-                | Hydrate (monitoredContext, hydrationConcurrency) -> LoadMode.Hydrated (categoryFilter, hydrationConcurrency, monitoredContext)
+        let start (log, storeLog) (sink : Propulsion.Streams.Default.Sink) categories
+            (indexStore, checkpoints, loadMode, startFromTail, batchSizeCutoff, tailSleepInterval, statsInterval) : Propulsion.Pipeline * (TimeSpan -> Task<unit>) option =
             let source =
                 DynamoStoreSource(
                     log, statsInterval,
                     indexStore, batchSizeCutoff, tailSleepInterval,
-                    checkpoints, sink, loadMode,
+                    checkpoints, sink, loadMode, categories = categories,
                     startFromTail = startFromTail, storeLog = storeLog)
             let source = source.Start()
             source, Some (fun propagationDelay -> source.Monitor.AwaitCompletion(propagationDelay, ignoreSubsequent = false))
@@ -91,43 +84,43 @@ module SourceConfig =
 #if esdb    
     module Esdb =
         open Propulsion.EventStoreDb
-        let start log (sink : Propulsion.Streams.Default.Sink) categoryFilter
-            (client, checkpoints, hydrateBodies, startFromTail, batchSize, tailSleepInterval, statsInterval) : Propulsion.Pipeline * (TimeSpan -> Task<unit>) option =
+        let start log (sink : Propulsion.Streams.Default.Sink) categories
+            (client, checkpoints, withData, startFromTail, batchSize, tailSleepInterval, statsInterval) : Propulsion.Pipeline * (TimeSpan -> Task<unit>) option =
             let source =
                 EventStoreSource(
                     log, statsInterval,
                     client, batchSize, tailSleepInterval,
-                    checkpoints, sink, categoryFilter, hydrateBodies = hydrateBodies, startFromTail = startFromTail)
+                    checkpoints, sink, categories, withData = withData, startFromTail = startFromTail)
             let source = source.Start()
             source, Some (fun propagationDelay -> source.Monitor.AwaitCompletion(propagationDelay, ignoreSubsequent = false))
 #endif            
 #if sss    
     module Sss =
         open Propulsion.SqlStreamStore
-        let start log (sink : Propulsion.Streams.Default.Sink) categoryFilter
-            (client, checkpoints, hydrateBodies, startFromTail, batchSize, tailSleepInterval, statsInterval) : Propulsion.Pipeline * (TimeSpan -> Task<unit>) option =
+        let start log (sink : Propulsion.Streams.Default.Sink) categories
+            (client, checkpoints, withData, startFromTail, batchSize, tailSleepInterval, statsInterval) : Propulsion.Pipeline * (TimeSpan -> Task<unit>) option =
             let source =
                 SqlStreamStoreSource(
                     log, statsInterval,
                     client, batchSize, tailSleepInterval,
-                    checkpoints, sink, categoryFilter, hydrateBodies = hydrateBodies, startFromTail = startFromTail)
+                    checkpoints, sink, categories, withData = withData, startFromTail = startFromTail)
             let source = source.Start()
             source, Some (fun propagationDelay -> source.Monitor.AwaitCompletion(propagationDelay, ignoreSubsequent = false))
 #endif            
-    let start (log, storeLog) sink categoryFilter : SourceConfig -> Propulsion.Pipeline * (TimeSpan -> Task<unit>) option = function
+    let start (log, storeLog) sink categories : SourceConfig -> Propulsion.Pipeline * (TimeSpan -> Task<unit>) option = function
 // #if cosmos    
         | SourceConfig.Cosmos (monitored, leases, checkpointConfig, tailSleepInterval) ->
-            Cosmos.start log sink categoryFilter (monitored, leases, checkpointConfig, tailSleepInterval)
+            Cosmos.start log sink categories (monitored, leases, checkpointConfig, tailSleepInterval)
 // #endif
 #if dynamo    
-        | SourceConfig.Dynamo (indexStore, checkpoints, loading, startFromTail, batchSizeCutoff, tailSleepInterval, statsInterval) ->
-            Dynamo.start (log, storeLog) sink categoryFilter (indexStore, checkpoints, loading, startFromTail, batchSizeCutoff, tailSleepInterval, statsInterval)
+        | SourceConfig.Dynamo (indexStore, checkpoints, loadMode, startFromTail, batchSizeCutoff, tailSleepInterval, statsInterval) ->
+            Dynamo.start (log, storeLog) sink categories (indexStore, checkpoints, loadMode, startFromTail, batchSizeCutoff, tailSleepInterval, statsInterval)
 #endif
 #if esdb    
-        | SourceConfig.Esdb (client, checkpoints, hydrateBodies, startFromTail, batchSize, tailSleepInterval, statsInterval) ->
-            Esdb.start log sink categoryFilter (client, checkpoints, hydrateBodies, startFromTail, batchSize, tailSleepInterval, statsInterval)
+        | SourceConfig.Esdb (client, checkpoints, withData, startFromTail, batchSize, tailSleepInterval, statsInterval) ->
+            Esdb.start log sink categories (client, checkpoints, withData, startFromTail, batchSize, tailSleepInterval, statsInterval)
 #endif
 #if sss    
-        | SourceConfig.Sss (client, checkpoints, hydrateBodies, startFromTail, batchSize, tailSleepInterval, statsInterval) ->
-            Sss.start log sink categoryFilter (client, checkpoints, hydrateBodies, startFromTail, batchSize, tailSleepInterval, statsInterval)
+        | SourceConfig.Sss (client, checkpoints, withData, startFromTail, batchSize, tailSleepInterval, statsInterval) ->
+            Sss.start log sink categories (client, checkpoints, withData, startFromTail, batchSize, tailSleepInterval, statsInterval)
 #endif