+++ propulsion-reactor/SourceConfig.fs
@@ -11,21 +11,21 @@ type SourceConfig =
         * tailSleepInterval : TimeSpan
     | Dynamo of indexStore : Equinox.DynamoStore.DynamoStoreClient
         * checkpoints : Propulsion.Feed.IFeedCheckpointStore
-        * loading : DynamoLoadModeConfig
+        * loading : Propulsion.DynamoStore.EventLoadMode
         * startFromTail : bool
         * batchSizeCutoff : int
         * tailSleepInterval : TimeSpan
         * statsInterval : TimeSpan
     | Esdb of client : EventStore.Client.EventStoreClient
         * checkpoints : Propulsion.Feed.IFeedCheckpointStore
-        * hydrateBodies : bool
+        * withData : bool
         * startFromTail : bool
         * batchSize : int
         * tailSleepInterval : TimeSpan
         * statsInterval : TimeSpan
     | Sss of client : SqlStreamStore.IStreamStore
         * checkpoints : Propulsion.Feed.IFeedCheckpointStore
-        * hydrateBodies : bool
+        * withData : bool
         * startFromTail : bool
         * batchSize : int
         * tailSleepInterval : TimeSpan
@@ -33,15 +33,13 @@ type SourceConfig =
 and [<NoEquality; NoComparison>] CosmosFeedConfig =
     | Ephemeral of processorName : string
     | Persistent of processorName : string * startFromTail : bool * maxItems : int option * lagFrequency : TimeSpan
-and [<NoEquality; NoComparison>] DynamoLoadModeConfig =
-    | Hydrate of monitoredContext : Equinox.DynamoStore.DynamoStoreContext * hydrationConcurrency : int
 
 module SourceConfig =
     module Cosmos =
         open Propulsion.CosmosStore
-        let start log (sink : Propulsion.Streams.Default.Sink) categoryFilter
+        let start log (sink : Propulsion.Streams.Default.Sink) categories
             (monitoredContainer, leasesContainer, checkpointConfig, tailSleepInterval) : Propulsion.Pipeline * (TimeSpan -> Task<unit>) option =
-            let parseFeedDoc = EquinoxSystemTextJsonParser.enumStreamEvents categoryFilter
+            let parseFeedDoc = EquinoxSystemTextJsonParser.enumStreamEvents (fun x -> Array.contains x categories) 
             let observer = CosmosStoreSource.CreateObserver(log, sink.StartIngester, Seq.collect parseFeedDoc)
             let source =
                 match checkpointConfig with
@@ -59,48 +57,45 @@ module SourceConfig =
             source, None
     module Dynamo =
         open Propulsion.DynamoStore
-        let start (log, storeLog) (sink : Propulsion.Streams.Default.Sink) categoryFilter
-            (indexStore, checkpoints, loadModeConfig, startFromTail, batchSizeCutoff, tailSleepInterval, statsInterval) : Propulsion.Pipeline * (TimeSpan -> Task<unit>) option =
-            let loadMode =
-                match loadModeConfig with
-                | Hydrate (monitoredContext, hydrationConcurrency) -> LoadMode.Hydrated (categoryFilter, hydrationConcurrency, monitoredContext)
+        let start (log, storeLog) (sink : Propulsion.Streams.Default.Sink) categories
+            (indexStore, checkpoints, loadMode, startFromTail, batchSizeCutoff, tailSleepInterval, statsInterval) : Propulsion.Pipeline * (TimeSpan -> Task<unit>) option =
             let source =
                 DynamoStoreSource(
                     log, statsInterval,
                     indexStore, batchSizeCutoff, tailSleepInterval,
-                    checkpoints, sink, loadMode,
+                    checkpoints, sink, loadMode, categories = categories,
                     startFromTail = startFromTail, storeLog = storeLog)
             let source = source.Start()
             source, Some (fun propagationDelay -> source.Monitor.AwaitCompletion(propagationDelay, ignoreSubsequent = false))
     module Esdb =
         open Propulsion.EventStoreDb
-        let start log (sink : Propulsion.Streams.Default.Sink) categoryFilter
-            (client, checkpoints, hydrateBodies, startFromTail, batchSize, tailSleepInterval, statsInterval) : Propulsion.Pipeline * (TimeSpan -> Task<unit>) option =
+        let start log (sink : Propulsion.Streams.Default.Sink) categories
+            (client, checkpoints, withData, startFromTail, batchSize, tailSleepInterval, statsInterval) : Propulsion.Pipeline * (TimeSpan -> Task<unit>) option =
             let source =
                 EventStoreSource(
                     log, statsInterval,
                     client, batchSize, tailSleepInterval,
-                    checkpoints, sink, categoryFilter, hydrateBodies = hydrateBodies, startFromTail = startFromTail)
+                    checkpoints, sink, (fun x -> Array.contains x categories), withData = withData, startFromTail = startFromTail)
             let source = source.Start()
             source, Some (fun propagationDelay -> source.Monitor.AwaitCompletion(propagationDelay, ignoreSubsequent = false))
     module Sss =
         open Propulsion.SqlStreamStore
-        let start log (sink : Propulsion.Streams.Default.Sink) categoryFilter
-            (client, checkpoints, hydrateBodies, startFromTail, batchSize, tailSleepInterval, statsInterval) : Propulsion.Pipeline * (TimeSpan -> Task<unit>) option =
+        let start log (sink : Propulsion.Streams.Default.Sink) categories
+            (client, checkpoints, withData, startFromTail, batchSize, tailSleepInterval, statsInterval) : Propulsion.Pipeline * (TimeSpan -> Task<unit>) option =
             let source =
                 SqlStreamStoreSource(
                     log, statsInterval,
                     client, batchSize, tailSleepInterval,
-                    checkpoints, sink, categoryFilter, hydrateBodies = hydrateBodies, startFromTail = startFromTail)
+                    checkpoints, sink, (fun x -> Array.contains x categories), withData = withData, startFromTail = startFromTail)
             let source = source.Start()
             source, Some (fun propagationDelay -> source.Monitor.AwaitCompletion(propagationDelay, ignoreSubsequent = false))
             
-    let start (log, storeLog) sink categoryFilter : SourceConfig -> Propulsion.Pipeline * (TimeSpan -> Task<unit>) option = function
+    let start (log, storeLog) sink categories : SourceConfig -> Propulsion.Pipeline * (TimeSpan -> Task<unit>) option = function
         | SourceConfig.Cosmos (monitored, leases, checkpointConfig, tailSleepInterval) ->
-            Cosmos.start log sink categoryFilter (monitored, leases, checkpointConfig, tailSleepInterval)
-        | SourceConfig.Dynamo (indexStore, checkpoints, loading, startFromTail, batchSizeCutoff, tailSleepInterval, statsInterval) ->
-            Dynamo.start (log, storeLog) sink categoryFilter (indexStore, checkpoints, loading, startFromTail, batchSizeCutoff, tailSleepInterval, statsInterval)
-        | SourceConfig.Esdb (client, checkpoints, hydrateBodies, startFromTail, batchSize, tailSleepInterval, statsInterval) ->
-            Esdb.start log sink categoryFilter (client, checkpoints, hydrateBodies, startFromTail, batchSize, tailSleepInterval, statsInterval)
-        | SourceConfig.Sss (client, checkpoints, hydrateBodies, startFromTail, batchSize, tailSleepInterval, statsInterval) ->
-            Sss.start log sink categoryFilter (client, checkpoints, hydrateBodies, startFromTail, batchSize, tailSleepInterval, statsInterval)
+            Cosmos.start log sink categories (monitored, leases, checkpointConfig, tailSleepInterval)
+        | SourceConfig.Dynamo (indexStore, checkpoints, loadMode, startFromTail, batchSizeCutoff, tailSleepInterval, statsInterval) ->
+            Dynamo.start (log, storeLog) sink categories (indexStore, checkpoints, loadMode, startFromTail, batchSizeCutoff, tailSleepInterval, statsInterval)
+        | SourceConfig.Esdb (client, checkpoints, withData, startFromTail, batchSize, tailSleepInterval, statsInterval) ->
+            Esdb.start log sink categories (client, checkpoints, withData, startFromTail, batchSize, tailSleepInterval, statsInterval)
+        | SourceConfig.Sss (client, checkpoints, withData, startFromTail, batchSize, tailSleepInterval, statsInterval) ->
+            Sss.start log sink categories (client, checkpoints, withData, startFromTail, batchSize, tailSleepInterval, statsInterval)